build: pin GitHub Actions
Pin GitHub Actions to their latest versions using hashes instead of tags
to prevent tampering. Tags are still added as a comment and both tags
and hashes are updated by Dependabot in subsequent pull requests.

Add `--frozen-lockfile` to `pnpm install`s to make it explicit that this
is what's happening in ci.